reimplemented DernieresMaraudes as a dropdown instead of ContextMixin
