+++ maraudes/menu.py
@@ -1,5 +1,6 @@
 #-*- coding:utf-8 -*-
 
+from django.utils.functional import cached_property
 from navbar import ApplicationMenu, DropDown
 
 # TODO:
@@ -8,9 +9,28 @@ from navbar import ApplicationMenu, DropDown
 # We shall find a way to do so, without creating problems
 # with the url resolver !
 # It cannot be in views.py, maybe when app is ready ?
-
 class DernieresMaraudes(DropDown):
-    pass
+    header = "Dernières maraudes"
+    
+    count = 4
+    @cached_property
+    def dernieres_maraudes(self):
+        """ Renvoie la liste des 'Maraude' passées et terminées """
+        #WARNING: Cached property has become inefficient here !
+        print('DEBUG_INFO: cached_property retrieves some data :')
+        from maraudes.models import Maraude
+        return Maraude.objects.get_past().filter(
+                                            heure_fin__isnull=False
+                                        ).order_by(
+                                            '-date'
+                                        )[:self.count]
+
+    def get_links(self):
+        return [
+            (   str(maraude), 
+                ('maraudes:details', {'kwargs': {'pk': maraude.pk}}),
+                None) for maraude in self.dernieres_maraudes ]
+
 
 
 
@@ -24,5 +44,7 @@ class MaraudesMenu(ApplicationMenu):
             ]
 
     def get_dropdowns(self, view):
-        pass
+        return [
+            DernieresMaraudes(),
+        ]
