+++ navbar/templates/navbar/navbar-menu.html
@@ -11,10 +11,9 @@
   </li>
 {% endfor %}
 {% for dropdown in dropdowns %}
-  <li class="dropdown app-menu"><a class="dropdown-toggle" data-toggle="dropdown" href="#">Dernières maraudes <b class="caret"></b></a>
-    <ul class="dropdown-menu">{% for m in dernieres_maraudes %}
-      <li><a href="{% url 'maraudes:details' m.pk %}" {% if maraude == m %}style="color:#fff;"{% endif %}>
-        <strong>{{ m }}</strong> <small>{{m.binome}} & {{m.referent}}</small></a>
+<li class="dropdown app-menu"><a class="dropdown-toggle" data-toggle="dropdown" href="#">{{ dropdown.header }} <b class="caret"></b></a>
+      <ul class="dropdown-menu">{% for link in dropdown.links %}
+	<li><a href="{{ link.href }}">{% if link.icon %}{% bootstrap_icon link.icon %}{% endif %} {{ link.text}}</a> 
       </li>{% endfor %}
     </ul>
   </li>
